Always get a new Vault token rather then renewing it
Because of the way we changed vault-sidekick to not panic when it fails
to renew a Vault token, we may end up in a situation where
vault-sidekick tries to renew an expired token (which results in a 403
response from Vault) leading to an endless loop.

This commit changes the renew logic to always get a new token rather
than trying to renew it.